pyproject.toml
[project]
name = "llm-tools"
version = "0.1.0"
description = "LLM-powered pipelines for CYOA narrative tooling"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.5.7",
    "jsonlines>=4.0.0",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "langchain-text-splitters>=1.1.2",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.2",
    "typer>=0.15.0",
]

[project.scripts]
llm-tools = "llm_tools.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
per-file-ignores."src/llm_tools/prompts/*" = ["E501"]
per-file-ignores."tests/conftest.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_tools"]
